chore: refactor component structure in scoped variables
